refactor: 独立长消息处理为longtext包
